recommend-service/pyproject.toml
[project]
name = "recommend-service"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "dannhanong", email = "thannong0512@gmail.com"},
]
dependencies = ["pymongo>=4.12.1", "scikit-learn>=1.6.1", "pandas>=2.2.3", "numpy>=2.2.5", "pyjwt>=2.10.1", "fastapi>=0.115.12", "uvicorn>=0.34.2", "pydantic-settings>=2.9.1", "python-dotenv>=1.1.0"]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
